+++ rs/src/control.rs
@@ -1,0 +1,22 @@
+#[cfg(feature = "serde")]
+use ::serde::{Deserialize, Serialize};
+
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub struct AbcHeader {
+  pub flags: u32,
+  pub name: String,
+}
+
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub struct Label {
+  pub frame: u32,
+  pub name: String,
+}
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub struct Scene {
+  pub offset: u32,
+  pub name: String,
+}